requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
markdown
django-filter
django-cors-headers
django-nose
coverage
requests
django-storages==1.11.1
gunicorn==20.1.0
boto3==1.17.58